Move models to db package
