+++ app/src/main/java/com/mbonetti/tmchallenge/db/models/Accessibility.kt
@@ -1,4 +1,4 @@
-package com.mbonetti.tmchallenge.models
+package com.mbonetti.tmchallenge.db.models
 
 data class Accessibility(
     val info: String
