+++ app/src/main/java/com/mbonetti/tmchallenge/db/models/City.kt
@@ -1,0 +1,5 @@
+package com.mbonetti.tmchallenge.db.models
+
+data class City(
+    val name: String
+)
