+++ app/src/main/java/com/mbonetti/tmchallenge/db/models/Embedded.kt
@@ -1,0 +1,5 @@
+package com.mbonetti.tmchallenge.db.models
+
+data class Embedded(
+    val events: List<com.mbonetti.tmchallenge.db.models.Event>
+)
