+++ app/src/main/java/com/mbonetti/tmchallenge/db/models/EmbeddedVenue.kt
@@ -1,0 +1,5 @@
+package com.mbonetti.tmchallenge.db.models
+
+data class EmbeddedVenue(
+    val venues: List<com.mbonetti.tmchallenge.db.models.Venue>
+)
