+++ app/src/main/java/com/mbonetti/tmchallenge/db/models/Promoter.kt
@@ -1,0 +1,7 @@
+package com.mbonetti.tmchallenge.db.models
+
+data class Promoter(
+    val id: String,
+    val name: String,
+    var description: String,
+)
