+++ app/src/main/java/com/mbonetti/tmchallenge/db/models/Public.kt
@@ -1,4 +1,4 @@
-package com.mbonetti.tmchallenge.models
+package com.mbonetti.tmchallenge.db.models
 
 data class Public(
     val endDateTime: String,
