--- app/src/main/java/com/mbonetti/tmchallenge/models/City.kt
@@ -1,5 +1,0 @@
-package com.mbonetti.tmchallenge.models
-
-data class City(
-    val name: String
-)
