--- app/src/main/java/com/mbonetti/tmchallenge/models/Dates.kt
@@ -1,8 +1,0 @@
-package com.mbonetti.tmchallenge.models
-
-import com.squareup.moshi.Json
-
-data class Dates(
-    @Json(name = "start")
-    val startDate: StartDate,
-)
