--- app/src/main/java/com/mbonetti/tmchallenge/models/Embedded.kt
@@ -1,5 +1,0 @@
-package com.mbonetti.tmchallenge.models
-
-data class Embedded(
-    val events: List<Event>
-)
