--- app/src/main/java/com/mbonetti/tmchallenge/models/EmbeddedVenue.kt
@@ -1,5 +1,0 @@
-package com.mbonetti.tmchallenge.models
-
-data class EmbeddedVenue(
-    val venues: List<Venue>
-)
