--- app/src/main/java/com/mbonetti/tmchallenge/models/EventImage.kt
@@ -1,9 +1,0 @@
-package com.mbonetti.tmchallenge.models
-
-data class EventImage(
-    val fallback: Boolean,
-    val height: Int,
-    val ratio: String,
-    val url: String,
-    val width: Int
-)
