--- app/src/main/java/com/mbonetti/tmchallenge/models/Promoter.kt
@@ -1,7 +1,0 @@
-package com.mbonetti.tmchallenge.models
-
-data class Promoter(
-    val description: String,
-    val id: String,
-    val name: String
-)
