--- app/src/main/java/com/mbonetti/tmchallenge/models/Sales.kt
@@ -1,6 +1,0 @@
-package com.mbonetti.tmchallenge.models
-
-data class Sales(
-    val presales: List<Presale>,
-    val `public`: Public
-)
