--- app/src/main/java/com/mbonetti/tmchallenge/models/Status.kt
@@ -1,5 +1,0 @@
-package com.mbonetti.tmchallenge.models
-
-data class Status(
-    val code: String
-)
